remove the cast

+++ 0x06-python-classes/103-magic_class.py
@@ -25,8 +25,8 @@ class MagicClass:
 
     def area(self):
         """return the area of of the radius"""
-        return float(self.__radius ** 2) * math.pi)
+        return (self.__radius ** 2) * math.pi)
 
     def circumference(self):
         """return the circumference of the radius"""
-        return float(2 * math.pi * self.__radius)
+        return (2 * math.pi * self.__radius)
